requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
sqlalchemy==2.0.37
sqlite-utils==3.38
python-dotenv==1.0.1
pytest==8.3.4
httpx==0.28.1
langgraph>=0.0.15
openai>=1.12.0
tavily-python>=0.1.9
tiktoken>=0.5.2

